refactor(engine): fold internal engine module into scan pipeline and tidy imports
