--- src/engine/mod.rs
@@ -1,2 +1,0 @@
-pub mod diagnostics;
-pub mod pipeline;
