+++ src/lib.rs
@@ -31,8 +31,6 @@ pub mod config;
 #[doc(hidden)]
 pub mod doctor;
 #[doc(hidden)]
-pub mod engine;
-#[doc(hidden)]
 pub mod explain;
 #[doc(hidden)]
 pub mod findings;
