+++ src/scan/scanner/contract_stage.rs
@@ -1,9 +1,38 @@
+//! Finding-contract validation as a scan pipeline stage.
+//!
+//! Runs after risk scoring to validate that every finding satisfies the finding
+//! contract, and turns any violations into scan diagnostics. Co-located with the
+//! other scanner stages (`finalize`, `summary`, …) because it is only ever run as
+//! part of a scan.
+
 use crate::findings::contract::{
     FindingContractReport, FindingContractViolation, FindingContractViolationKind,
+    validate_findings_contract,
 };
+use crate::findings::types::Finding;
 use crate::scan::types::ScanDiagnostic;
+use std::time::Instant;
 
-pub fn finding_contract_diagnostics(report: &FindingContractReport) -> Vec<ScanDiagnostic> {
+pub struct FindingContractValidationStage {
+    pub elapsed_us: u64,
+    pub report: FindingContractReport,
+    pub diagnostics: Vec<ScanDiagnostic>,
+}
+
+pub fn validate_finding_contract_stage(findings: &[Finding]) -> FindingContractValidationStage {
+    let start = Instant::now();
+    let report = validate_findings_contract(findings);
+    let elapsed_us = start.elapsed().as_micros() as u64;
+    let diagnostics = finding_contract_diagnostics(&report);
+
+    FindingContractValidationStage {
+        elapsed_us,
+        report,
+        diagnostics,
+    }
+}
+
+fn finding_contract_diagnostics(report: &FindingContractReport) -> Vec<ScanDiagnostic> {
     if report.violations.is_empty() {
         return Vec::new();
     }
